+++ module/Formation/src/Formation/Controller/InscriptionControllerFactory.php
@@ -3,6 +3,8 @@
 namespace Formation\Controller;
 
 use Fichier\Service\Fichier\FichierStorageService;
+use Formation\Service\Exporter\Attestation\AttestationExporter;
+use Formation\Service\Exporter\Convocation\ConvocationExporter;
 use Formation\Service\Notification\FormationNotificationFactory;
 use Formation\Service\Session\SessionService;
 use Psr\Container\ContainerExceptionInterface;
@@ -39,6 +41,7 @@ class InscriptionControllerFactory {
          * @var PresenceService $presenceService
          * @var SessionService $sessionService
          * @var StructureDocumentService $structureDocumentService
+         * @var AttestationExporter $attestationExporter
          */
         $entityManager = $container->get('doctrine.entitymanager.orm_default');
         $doctorantService = $container->get(DoctorantService::class);
@@ -50,6 +53,8 @@ class InscriptionControllerFactory {
         $presenceService = $container->get(PresenceService::class);
         $sessionService = $container->get(SessionService::class);
         $structureDocumentService = $container->get(StructureDocumentService::class);
+        $attestationExporter = $container->get(AttestationExporter::class);
+        $convocationExporter = $container->get(ConvocationExporter::class);
 
         /* @var $renderer PhpRenderer */
         $renderer = $container->get('ViewRenderer');
@@ -68,6 +73,8 @@ class InscriptionControllerFactory {
         $controller->setStructureDocumentService($structureDocumentService);
         /** forms *****************************************************************************************************/
         /** autres*****************************************************************************************************/
+        $controller->setAttestationExporter($attestationExporter);
+        $controller->setConvocationExporter($convocationExporter);
         $controller->setRenderer($renderer);
 
         /** @var \Formation\Service\Notification\FormationNotificationFactory $formationNotificationFactory */