+++ module/Formation/src/Formation/Controller/Recherche/ModuleRechercheControllerFactory.php
@@ -4,6 +4,7 @@ namespace Formation\Controller\Recherche;
 
 use Formation\Service\Module\Search\ModuleSearchService;
 use Psr\Container\ContainerInterface;
+use UnicaenParametre\Service\Parametre\ParametreService;
 
 class ModuleRechercheControllerFactory
 {
@@ -13,11 +14,17 @@ class ModuleRechercheControllerFactory
      */
     public function __invoke(ContainerInterface $container): ModuleRechercheController
     {
-        /** @var \Formation\Service\Module\Search\ModuleSearchService $searchService */
+        /**
+         * @var \Formation\Service\Module\Search\ModuleSearchService $searchService
+         * @var ParametreService $parametreService
+         */
         $searchService = $container->get(ModuleSearchService::class);
+        $parametreService = $container->get(ParametreService::class);
+
 
         $controller = new ModuleRechercheController();
         $controller->setSearchService($searchService);
+        $controller->setParametreService($parametreService);
 
         return $controller;
     }