+++ module/Formation/src/Formation/Provider/Parametre/FormationParametres.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Formation\Provider\Parametre;
+
+class FormationParametres
+{
+    const CATEGORIE = 'FORMATION';
+
+    const DELAI_ENQUETE = 'DELAI_ENQUETE';
+}