+++ module/Formation/src/Formation/Service/Exporter/Convocation/ConvocationExporterAwareTrait.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Formation\Service\Exporter\Convocation;
+
+trait ConvocationExporterAwareTrait {
+
+    private ConvocationExporter $convocationExporter;
+
+    public function getConvocationExporter(): ConvocationExporter
+    {
+        return $this->convocationExporter;
+    }
+
+    public function setConvocationExporter(ConvocationExporter $convocationExporter): void
+    {
+        $this->convocationExporter = $convocationExporter;
+    }
+
+}