+++ module/Soutenance/src/Soutenance/Service/Url/UrlServiceFactory.php
@@ -6,6 +6,7 @@ use Interop\Container\ContainerInterface;
 use Laminas\View\Renderer\PhpRenderer;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
+use Soutenance\Service\Membre\MembreService;
 
 class UrlServiceFactory {
 
@@ -17,11 +18,16 @@ class UrlServiceFactory {
      */
     public function __invoke(ContainerInterface $container) : UrlService
     {
-        /* @var PhpRenderer $renderer  */
+        /**
+         * @var PhpRenderer $renderer
+         * @var MembreService $membreService
+         */
         $renderer = $container->get('ViewRenderer');
+        $membreService = $container->get(MembreService::class);
 
         $service = new UrlService();
         $service->setRenderer($renderer);
+        $service->setMembreService($membreService);
         return $service;
     }
 }